refactor: 콘솔 입출력을 담당하는 인터페이스를 추가 및 TextIO로 구현
