+++ src/main/java/com/programmers/voucher/global/io/Console.java
@@ -1,15 +1,4 @@
 package com.programmers.voucher.global.io;
 
-import com.programmers.voucher.domain.voucher.dto.request.VoucherCreateRequest;
-
-public interface Console {
-    ConsoleCommandType inputInitialCommand();
-
-    void printCommandSet();
-
-    VoucherCreateRequest inputVoucherCreateInfo();
-
-    void print(String result);
-
-    void exit();
+public interface Console extends ConsoleInput, ConsoleOutput {
 }
