+++ src/main/java/com/programmers/voucher/global/io/ConsoleInput.java
@@ -1,0 +1,9 @@
+package com.programmers.voucher.global.io;
+
+import com.programmers.voucher.domain.voucher.dto.request.VoucherCreateRequest;
+
+public interface ConsoleInput {
+    ConsoleCommandType inputInitialCommand();
+
+    VoucherCreateRequest inputVoucherCreateInfo();
+}
