+++ src/main/java/com/programmers/voucher/global/io/ConsoleOutput.java
@@ -1,0 +1,9 @@
+package com.programmers.voucher.global.io;
+
+public interface ConsoleOutput {
+    void printCommandSet();
+
+    void print(String result);
+
+    void exit();
+}
